Revert "Unconditionally clean out/"
This reverts commit 5562302e414b3a84dc00c0612b5969d392b1fa7f.

+++ misc_functions.sh
@@ -114,10 +114,10 @@ function restore_saved_build_state {
 
 function clean_out {
     cd $BUILD_TOP/
-#    if [ "x${CLEAN_TARGET_OUT}" != "x" ] && [ ${CLEAN_TARGET_OUT} -eq 1 ]; then
-    echoText "Cleaning build dir..."
-    rm -rf out
-#    fi
+    if [ "x${CLEAN_TARGET_OUT}" != "x" ] && [ ${CLEAN_TARGET_OUT} -eq 1 ]; then
+        echoText "Cleaning build dir..."
+        rm -rf out
+    fi
 }
 
 function clean_state {
